requirements.txt
aiogram==3.4.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
python-pptx==0.6.23
aiofiles==23.2.1
aiohttp==3.9.1
